+++ examples/04-logger/src/app/app.module.ts
@@ -1,12 +1,10 @@
-import { RootModule, LoggerConfig } from '@ditsmod/core';
+import { RootModule, LoggerConfig, Providers } from '@ditsmod/core';
 import { RouterModule } from '@ditsmod/router';
 
 import { BunyanModule } from './modules/bunyan/bunyan.module';
 import { PinoModule } from './modules/pino/pino.module';
 import { SomeModule } from './modules/some/some.module';
 import { WinstonModule } from './modules/winston/winston.module';
-
-const loggerConfig = new LoggerConfig('info');
 
 @RootModule({
   imports: [
@@ -17,7 +15,7 @@ const loggerConfig = new LoggerConfig('info');
     { path: '', module: BunyanModule },
   ],
   providersPerApp: [
-    { provide: LoggerConfig, useValue: loggerConfig }
+    ...new Providers().useValue(LoggerConfig, new LoggerConfig('info')),
   ]
 })
 export class AppModule {}