+++ examples/11-override-core-log-messages/src/app/modules/other/other-log-mediator.ts
@@ -1,4 +1,4 @@
-import { LogFilter } from '@ditsmod/core';
+import { MsgLogFilter } from '@ditsmod/core';
 
 import { SomeLogMediator } from '../some/some-log-mediator';
 
@@ -8,8 +8,8 @@ export class OtherLogMediator extends SomeLogMediator {
    */
   override someNewMethod(self: object, additionalArgument: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('info', logFilter, `OtherLogMediator with overrided someNewMethod says: "${additionalArgument}"`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('info', msgLogFilter, `OtherLogMediator with overrided someNewMethod says: "${additionalArgument}"`);
   }
 }