+++ packages/core/src/services/log-mediator.ts
@@ -1,7 +1,7 @@
 import { Injectable, Optional } from '@ts-stack/di';
 
 import { isInjectionToken } from '../utils/type-guards';
-import { Logger, LogLevel } from '../types/logger';
+import { Logger, LoggerConfig, LogLevel } from '../types/logger';
 import { GlobalProviders, ImportObj } from '../types/metadata-per-mod';
 import { Extension, ExtensionsGroupToken, ModuleType, ModuleWithParams, ServiceProvider } from '../types/mix';
 import { getImportedTokens } from '../utils/get-imports';
@@ -18,15 +18,24 @@ export class LogFilter {
   classesNames?: string[];
   tags?: string[];
 }
+/**
+ * Uses by LogMediator.
+ */
+export class MsgLogFilter {
+  className?: string;
+  tags?: string[];
+}
 
 /**
  * Default type for Log buffer.
  */
 export interface LogItem {
+  moduleName: string;
   date: Date;
-  logFilter: LogFilter;
-  currentLevel: LogLevel;
-  messageLevel: LogLevel;
+  msgLogFilter: MsgLogFilter;
+  loggerLogFilter: LogFilter;
+  loggerLevel: LogLevel;
+  msgLevel: LogLevel;
   msg: string;
   logger: Logger;
 }
@@ -40,6 +49,7 @@ export interface LogItem {
  */
 @Injectable()
 export class LogMediator {
+  moduleName = 'AppModule';
   /**
    * If `bufferLogs === true` then all messages will be buffered.
    *
@@ -70,25 +80,30 @@ export class LogMediator {
   constructor(
     protected logManager: LogManager,
     @Optional() protected _logger: Logger = new ConsoleLogger(),
-    @Optional() protected logFilter: LogFilter = new LogFilter()
+    @Optional() protected logFilter: LogFilter = new LogFilter(),
+    @Optional() protected loggerConfig: LoggerConfig = new LoggerConfig(),
   ) {}
 
   getLogManager() {
     return this.logManager;
   }
 
-  protected setLog<T extends LogFilter>(level: LogLevel, logFilter: T, msg: any) {
+  protected setLog<T extends MsgLogFilter>(msgLevel: LogLevel, msgLogFilter: T, msg: any) {
     if (this.logManager.bufferLogs) {
+      const loggerLevel: LogLevel = typeof this._logger.getLevel == 'function' ? this._logger.getLevel() : this.loggerConfig.level;
+
       this.logManager.buffer.push({
+        moduleName: this.moduleName,
         logger: this._logger,
-        currentLevel: this._logger.config?.level || 'info',
-        logFilter,
+        loggerLevel,
+        loggerLogFilter: this.logFilter || new LogFilter(),
+        msgLevel,
+        msgLogFilter,
         date: new Date(),
-        messageLevel: level,
         msg,
       });
     } else {
-      this.logger.log(level, msg);
+      this.logger.log(msgLevel, msg);
     }
   }
 
@@ -99,14 +114,13 @@ export class LogMediator {
     const { buffer } = this.logManager;
     if (typeof (global as any).it != 'function') {
       // This is not a test mode.
-      const logFilter = this.logFilter;
       let filteredBuffer = buffer;
-      filteredBuffer = this.filterLogs(buffer, logFilter);
+      filteredBuffer = this.filterLogs(buffer);
       filteredBuffer.forEach((logItem) => {
         // const dateTime = log.date.toLocaleString();
-        const partMsg = logItem.logFilter.tags ? ` (Tags: ${logItem.logFilter.tags.join(', ')})` : '';
+        const partMsg = logItem.msgLogFilter.tags ? ` (Tags: ${logItem.msgLogFilter.tags.join(', ')})` : '';
         const msg = `${logItem.msg}${partMsg}`;
-        logItem.logger.setLevel(logItem.currentLevel);
+        logItem.logger.setLevel(logItem.loggerLevel);
 
         if (!logItem.logger.log) {
           const loggerName = logItem.logger.constructor.name;
@@ -117,7 +131,7 @@ export class LogMediator {
             console.error(msg0, msg);
           }
         } else {
-          logItem.logger.log.call(logItem.logger, logItem.messageLevel, msg);
+          logItem.logger.log.call(logItem.logger, logItem.msgLevel, msg);
         }
       });
     }
@@ -125,24 +139,41 @@ export class LogMediator {
     buffer.splice(0);
   }
 
-  protected filterLogs<T extends LogFilter>(buffer: LogItem[], outputConfig?: T) {
-    outputConfig = outputConfig || {} as T;
-    return buffer.filter((item) => {
-      const inputConfig = item.logFilter;
+  protected filterLogs(buffer: LogItem[]) {
+    const uniqFilters = new Map<LogFilter, string>();
+
+    const filteredBuffer = buffer.filter((item) => {
+      const { msgLogFilter, loggerLogFilter, moduleName } = item;
+      uniqFilters.set(loggerLogFilter, moduleName);
       let hasTags: boolean | undefined = true;
       let hasModuleName: boolean | undefined = true;
       let hasClassName: boolean | undefined = true;
-      if (outputConfig!.modulesNames) {
-        hasModuleName = inputConfig.modulesNames?.some((modulesName) => outputConfig!.modulesNames?.includes(modulesName));
+      if (loggerLogFilter!.modulesNames) {
+        hasModuleName = loggerLogFilter!.modulesNames?.includes(moduleName);
       }
-      if (outputConfig!.classesNames) {
-        hasClassName = inputConfig.classesNames?.some((className) => outputConfig!.classesNames?.includes(className));
+      if (loggerLogFilter!.classesNames) {
+        hasClassName = loggerLogFilter!.classesNames?.includes(msgLogFilter.className || '');
       }
-      if (outputConfig!.tags) {
-        hasTags = inputConfig.tags?.some((tag) => outputConfig!.tags?.includes(tag));
+      if (loggerLogFilter!.tags) {
+        hasTags = msgLogFilter.tags?.some((tag) => loggerLogFilter!.tags?.includes(tag));
       }
       return hasModuleName && hasClassName && hasTags;
     });
+
+    if (uniqFilters.size > 1) {
+      this.detectedDifferentLogFilters(uniqFilters);
+    }
+
+    return filteredBuffer;
+  }
+
+  protected detectedDifferentLogFilters(uniqFilters: Map<LogFilter, string>) {
+    const filtersStr: string[] = [];
+    uniqFilters.forEach((moduleName, filter) => {
+      filtersStr.push(`${moduleName} ${JSON.stringify(filter)}`)
+    });
+
+    this.logger.log.call(this.logger, 'warn', `LogMediator: detected ${uniqFilters.size} different LogFilters: ${filtersStr.join(', ')}`);
   }
 
   /**
@@ -150,11 +181,11 @@ export class LogMediator {
    */
   moduleAlreadyImported(self: object, inputModule: ModuleType | ModuleWithParams, targetModuleId: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const inputModuleId = getModuleName(inputModule);
     const msg = `${className}: "${inputModuleId}" has already been imported into "${targetModuleId}".`;
-    this.setLog('warn', logFilter, msg);
+    this.setLog('warn', msgLogFilter, msg);
   }
 
   /**
@@ -162,9 +193,9 @@ export class LogMediator {
    */
   serverListen(self: object, serverName: string, host: string, port: number) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('info', logFilter, `${className}: ${serverName} is running at http://${host}:${port}.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('info', msgLogFilter, `${className}: ${serverName} is running at http://${host}:${port}.`);
   }
 
   /**
@@ -172,9 +203,9 @@ export class LogMediator {
    */
   startReinitApp(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('debug', logFilter, `${className}: start reinit the application.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('debug', msgLogFilter, `${className}: start reinit the application.`);
   }
 
   /**
@@ -182,9 +213,9 @@ export class LogMediator {
    */
   skippingAutocommitModulesConfig(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('warn', logFilter, `${className}: skipping autocommit of changes for config of moduleManager.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('warn', msgLogFilter, `${className}: skipping autocommit of changes for config of moduleManager.`);
   }
 
   /**
@@ -192,9 +223,9 @@ export class LogMediator {
    */
   finishReinitApp(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('debug', logFilter, `${className}: finished reinit the application.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('debug', msgLogFilter, `${className}: finished reinit the application.`);
   }
 
   /**
@@ -202,9 +233,9 @@ export class LogMediator {
    */
   printReinitError(self: object, err: any) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('error', logFilter, err);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('error', msgLogFilter, err);
   }
 
   /**
@@ -212,10 +243,10 @@ export class LogMediator {
    */
   startRollbackModuleConfigChanges(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const msg = `${className}: start rollback of changes for config of moduleManager during reinit the application.`;
-    this.setLog('debug', logFilter, msg);
+    this.setLog('debug', msgLogFilter, msg);
   }
 
   /**
@@ -223,10 +254,10 @@ export class LogMediator {
    */
   successfulRollbackModuleConfigChanges(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const msg = `${className}: successful rollback of changes for config of moduleManager during reinit the application.`;
-    this.setLog('debug', logFilter, msg);
+    this.setLog('debug', msgLogFilter, msg);
   }
 
   /**
@@ -234,10 +265,10 @@ export class LogMediator {
    */
   successfulAddedModuleToImport(self: object, inputModule: ModuleType | ModuleWithParams, targetMetaName: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const inputModuleName = getModuleName(inputModule);
-    this.setLog('debug', logFilter, `${className}: successful added "${inputModuleName}" to "${targetMetaName}".`);
+    this.setLog('debug', msgLogFilter, `${className}: successful added "${inputModuleName}" to "${targetMetaName}".`);
   }
 
   /**
@@ -245,9 +276,9 @@ export class LogMediator {
    */
   moduleNotFound(self: object, moduleId: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('warn', logFilter, `${className}: module with ID "${moduleId}" not found.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('warn', msgLogFilter, `${className}: module with ID "${moduleId}" not found.`);
   }
 
   /**
@@ -255,19 +286,19 @@ export class LogMediator {
    */
   moduleSuccessfulRemoved(self: object, inputMetaName: string, targetMetaName: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('debug', logFilter, `${className}: ${inputMetaName} successful removed from ${targetMetaName}.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('debug', msgLogFilter, `${className}: ${inputMetaName} successful removed from ${targetMetaName}.`);
   }
 
   /**
    * `${moduleName} has ID: "${moduleId}".`
    */
-  moduleHasId(self: object, moduleName: string, moduleId: string) {
+  moduleHasId(self: object, moduleId: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('trace', logFilter, `${className}: ${moduleName} has ID: "${moduleId}".`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('trace', msgLogFilter, `${className}: ${this.moduleName} has ID: "${moduleId}".`);
   }
 
   /**
@@ -277,16 +308,16 @@ export class LogMediator {
    */
   printGlobalProviders(self: object, globalProviders: GlobalProviders) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const globalProvidersPerMod = this.getProvidersNames(globalProviders.importedProvidersPerMod);
     const globalProvidersPerRou = this.getProvidersNames(globalProviders.importedProvidersPerRou);
     const globalProvidersPerReq = this.getProvidersNames(globalProviders.importedProvidersPerReq);
     const prefix = `${className}: global providers per a`;
-    this.setLog('debug', logFilter, `${className}: global providers are collected.`);
-    this.setLog('trace', logFilter, `${prefix} module: [${globalProvidersPerMod}]`);
-    this.setLog('trace', logFilter, `${prefix} route: [${globalProvidersPerRou}]`);
-    this.setLog('trace', logFilter, `${prefix} request: [${globalProvidersPerReq}]`);
+    this.setLog('debug', msgLogFilter, `${className}: global providers are collected.`);
+    this.setLog('trace', msgLogFilter, `${prefix} module: [${globalProvidersPerMod}]`);
+    this.setLog('trace', msgLogFilter, `${prefix} route: [${globalProvidersPerRou}]`);
+    this.setLog('trace', msgLogFilter, `${prefix} request: [${globalProvidersPerReq}]`);
   }
 
   protected getProvidersNames(providersMap: Map<any, ImportObj<ServiceProvider>>) {
@@ -296,24 +327,22 @@ export class LogMediator {
   /**
    * ================== ModuleName ======================.
    */
-  startExtensionsModuleInit(self: object, moduleName: string) {
+  startExtensionsModuleInit(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.modulesNames = [moduleName];
-    logFilter.classesNames = [className];
-    this.setLog('debug', logFilter, `${className}: ${'='.repeat(20)} ${moduleName} ${'='.repeat(20)}`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('debug', msgLogFilter, `${className}: ${'='.repeat(20)} ${this.moduleName} ${'='.repeat(20)}`);
   }
 
   /**
    * `${tokenName} start init.`
    */
-  startExtensionsGroupInit(self: object, moduleName: string, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
+  startExtensionsGroupInit(self: object, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.modulesNames = [moduleName];
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const path = this.getExtentionPath(unfinishedInit);
-    this.setLog('trace', logFilter, `${className}: ${path}: start init.`);
+    this.setLog('trace', msgLogFilter, `${className}: ${path}: start init.`);
   }
 
   protected getExtentionPath(unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
@@ -329,13 +358,12 @@ export class LogMediator {
   /**
    * `finish init ${tokenName}.`
    */
-  finishExtensionsGroupInit(self: object, moduleName: string, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
+  finishExtensionsGroupInit(self: object, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.modulesNames = [moduleName];
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const path = this.getExtentionPath(unfinishedInit);
-    this.setLog('trace', logFilter, `${className}: ${path}: finish init.`);
+    this.setLog('trace', msgLogFilter, `${className}: ${path}: finish init.`);
   }
 
   /**
@@ -343,10 +371,10 @@ export class LogMediator {
    */
   noExtensionsFound(self: object, groupToken: any) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const tokenName = getProviderName(groupToken);
-    this.setLog('trace', logFilter, `${className}: for ${tokenName} no extensions found.`);
+    this.setLog('trace', msgLogFilter, `${className}: for ${tokenName} no extensions found.`);
   }
 
   /**
@@ -354,10 +382,10 @@ export class LogMediator {
    */
   startInitExtension(self: object, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const path = this.getExtentionPath(unfinishedInit);
-    this.setLog('trace', logFilter, `${className}: ${path}: start init.`);
+    this.setLog('trace', msgLogFilter, `${className}: ${path}: start init.`);
   }
 
   /**
@@ -365,23 +393,22 @@ export class LogMediator {
    */
   finishInitExtension(self: object, unfinishedInit: Set<Extension<any> | ExtensionsGroupToken<any>>, data: any) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const path = this.getExtentionPath(unfinishedInit);
     const withSomeValue = data === undefined ? ', no value returned' : ', returned some value';
-    this.setLog('trace', logFilter, `${className}: ${path}: finish init${withSomeValue}.`);
+    this.setLog('trace', msgLogFilter, `${className}: ${path}: finish init${withSomeValue}.`);
   }
 
   /**
    * `total inited ${extensionsNum} extensions: ${extensionsNames}.`
    */
-  totalInitedExtensions(self: object, moduleName: string, extensionsNum: number, extensionsNames: string) {
+  totalInitedExtensions(self: object, extensionsNum: number, extensionsNames: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.modulesNames = [moduleName];
-    logFilter.classesNames = [className];
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
     const msg = `${className}: total inited ${extensionsNum} extensions: ${extensionsNames}.`;
-    this.setLog('debug', logFilter, msg);
+    this.setLog('debug', msgLogFilter, msg);
   }
 
   /**
@@ -389,9 +416,9 @@ export class LogMediator {
    */
   controllerHasError(self: object, err: any) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('error', logFilter, err);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('error', msgLogFilter, err);
   }
 
   /**
@@ -399,9 +426,9 @@ export class LogMediator {
    */
   internalServerError(self: object, err: any) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('error', logFilter, err);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('error', msgLogFilter, err);
   }
 
   /**
@@ -409,9 +436,9 @@ export class LogMediator {
    */
   youCannotActivateRoute(self: object, httpMethod: string, url: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('debug', logFilter, `${className}: can not activate the route with URL: ${httpMethod} ${url}.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    this.setLog('debug', msgLogFilter, `${className}: can not activate the route with URL: ${httpMethod} ${url}.`);
   }
 
   /**
@@ -419,19 +446,20 @@ export class LogMediator {
    */
   noRoutes(self: object) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    this.setLog('warn', logFilter, `${className}: the application has no routes.`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    msgLogFilter.tags = ['route'];
+    this.setLog('warn', msgLogFilter, `${className}: the application has no routes.`);
   }
 
   /**
    * `setted route ${httpMethod} "/${path}"`.
    */
-  printRoute(self: object, moduleName: string, httpMethod: string, path: string) {
+  printRoute(self: object, httpMethod: string, path: string) {
     const className = self.constructor.name;
-    const logFilter = new LogFilter();
-    logFilter.classesNames = [className];
-    logFilter.modulesNames = [moduleName];
-    this.setLog('debug', logFilter, `${className}: setted route ${httpMethod} "/${path}".`);
+    const msgLogFilter = new MsgLogFilter();
+    msgLogFilter.className = className;
+    msgLogFilter.tags = ['route'];
+    this.setLog('debug', msgLogFilter, `${className}: setted route ${httpMethod} "/${path}".`);
   }
 }